requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
openai>=1.0.0
python-dotenv
easyocr
Pillow
PyMuPDF
numpy
python-docx
redis
motor
pymongo
PyJWT
cryptography